using std::bitset as key for styleParamCache

+++ core/src/scene/sceneLayer.cpp
@@ -2,7 +2,7 @@
 
 namespace Tangram {
 
-    uint32_t SceneLayer::s_layerCount = 0;
+    uint8_t SceneLayer::s_layerCount = 0;
 
     SceneLayer::SceneLayer(const std::vector<std::shared_ptr<SceneLayer>>&& _subLayers, const StyleParamMap&& _styleParamMap, const std::string _name, Filter* _filter) :
         m_subLayers(std::move(_subLayers)), m_styleParams(std::move(_styleParamMap)), m_name(_name), m_filter(_filter) {
